Factor out handler setup into a common module
